fix(android): Fix clicking on element out of parent bounds
