+++ src/Uno.UI/UI/Xaml/Controls/TextBlock/TextBlock.cs
@@ -988,6 +988,8 @@ namespace Windows.UI.Xaml.Controls
 
 		internal override bool CanHaveChildren() => true;
 
+		internal override bool IsViewHit() => true;
+
 		public new bool Focus(FocusState value) => base.Focus(value);
 
 		internal override bool IsFocusable =>
