+++ src/Uno.UI/UI/Xaml/Controls/TextBlock/TextBlock.netstd.cs
@@ -6,6 +6,5 @@ namespace Windows.UI.Xaml.Controls
 {
 	partial class TextBlock
 	{
-		internal override bool IsViewHit() => Text != null || base.IsViewHit();
 	}
 }
